+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/test/TestCommand.java
@@ -3,11 +3,16 @@ package ru.vyarus.dropwizard.guice.test;
 import io.dropwizard.core.Application;
 import io.dropwizard.core.Configuration;
 import io.dropwizard.core.cli.EnvironmentCommand;
+import io.dropwizard.core.setup.Bootstrap;
 import io.dropwizard.core.setup.Environment;
 import net.sourceforge.argparse4j.inf.Namespace;
 import org.eclipse.jetty.util.component.ContainerLifeCycle;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import ru.vyarus.dropwizard.guice.test.util.ConfigModifier;
+import ru.vyarus.dropwizard.guice.test.util.ConfigOverrideUtils;
+
+import java.util.List;
 
 /**
  * Lightweight variation of server command for testing purposes.
@@ -22,6 +27,7 @@ public class TestCommand<C extends Configuration> extends EnvironmentCommand<C> 
     private final Logger logger = LoggerFactory.getLogger(TestCommand.class);
     private final Class<C> configurationClass;
     private final boolean simulateManaged;
+    private final List<ConfigModifier<C>> modifiers;
     private ContainerLifeCycle container;
 
     public TestCommand(final Application<C> application) {
@@ -29,10 +35,27 @@ public class TestCommand<C extends Configuration> extends EnvironmentCommand<C> 
     }
 
     public TestCommand(final Application<C> application, final boolean simulateManaged) {
+        this(application, simulateManaged, null);
+    }
+
+    public TestCommand(final Application<C> application, final boolean simulateManaged,
+                       final List<ConfigModifier<C>> modifiers) {
         super(application, "guicey-test", "Specific command to run guice context without jetty server");
         cleanupAsynchronously();
         configurationClass = application.getConfigurationClass();
         this.simulateManaged = simulateManaged;
+        this.modifiers = modifiers;
+    }
+
+    @Override
+    protected void run(final Bootstrap<C> bootstrap,
+                       final Namespace namespace,
+                       final C configuration) throws Exception {
+        // at this point only logging configuration performed
+        if (modifiers != null) {
+            ConfigOverrideUtils.runModifiers(configuration, modifiers);
+        }
+        super.run(bootstrap, namespace, configuration);
     }
 
     @Override